pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anomaliz"
version = "0.1.0"
description = "Time-series anomaly detection service"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scikit-learn>=1.4",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "tensorflow-cpu>=2.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["anomaliz*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
